requirements.txt
pydantic
bcc
pyelftools